+++ backend/canisters/group/c2c_client/src/lib.rs
@@ -1,4 +1,4 @@
-use canister_client::{generate_c2c_call, generate_candid_c2c_call};
+use canister_client::generate_c2c_call;
 use group_canister::*;
 
 // Queries
@@ -11,10 +11,6 @@ generate_c2c_call!(c2c_name_and_members);
 generate_c2c_call!(public_summary);
 generate_c2c_call!(summary);
 generate_c2c_call!(summary_updates);
-
-generate_candid_c2c_call!(events);
-generate_candid_c2c_call!(invite_code);
-generate_candid_c2c_call!(selected_initial);
 
 // Updates
 generate_c2c_call!(c2c_delete_group);
@@ -33,8 +29,7 @@ generate_c2c_call!(c2c_tip_message);
 generate_c2c_call!(c2c_unfreeze_group);
 generate_c2c_call!(c2c_update_proposals);
 generate_c2c_call!(c2c_update_user_principal);
+generate_c2c_call!(change_role);
 generate_c2c_call!(delete_messages);
-
-generate_candid_c2c_call!(change_role);
-generate_candid_c2c_call!(send_message_v2);
-generate_candid_c2c_call!(update_group_v2);
+generate_c2c_call!(send_message_v2);
+generate_c2c_call!(update_group_v2);