+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -8,7 +8,6 @@ use ic_cdk::post_upgrade;
 use instruction_counts_log::InstructionCountFunctionId;
 use stable_memory::get_reader;
 use tracing::info;
-use utils::env::Environment;
 
 #[post_upgrade]
 #[trace]
@@ -18,14 +17,12 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
-
-    let env = init_env(data.rng_seed);
-    data.chat.events.remove_spurious_video_call_in_progress(env.now());
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
+    let env = init_env(data.rng_seed);
     init_state(env, data, args.wasm_version);
 
     info!(version = %args.wasm_version, "Post-upgrade complete");